Beaglebone: Add lcd cape card detection
This patch reads the eeprom info from LCD cape and initializes the LCD cape
LCD , keypad and touch screen.

By default setup DVI display if LCD daughter card is not detected.
Display needs to be initialized even if display daughter card is not found so as
to enable framebuffer driver which is needed for successful Android bootup

Signed-off-by: Pankaj Bharadiya <pankaj.bharadiya@ti.com>