usb: misc: introduce OTG & EH Test Driver
This patch introduces support for test modes as
defined by the USB OTG & Embedded host Specification.

Note that we _must_ be able to suspend & resume ports
as we wish, so this driver depends on USB_SUSPEND and
makes use (well, somewhat abuses) usb autopm features
to achieve what we want.

Signed-off-by: Felipe Balbi <balbi@ti.com>
Signed-off-by: Ravi Babu <ravibabu@ti.com>