ARM: OMAP: AM33XX: Disable OPP120 & OPPTURBO if BeableBone is USB powered
This patch adds support for disabling OPP120 and OPPTURBO if the
BeagleBone is powered with USB.

Maximum current provided by the USB port is 500mA which is not sufficient
when operating @OPP120 and OPPTURBO. If OPP120 and OPPTURBO are not disabled
and on-demand cpufreq governor is used this will lead to unpredictable
behavior. Based on the maximum current requirements for OPP100 as listed in
the AM335x Datasheet there might be issues with OPP100 also but for now
we just disable OPP120 and OPPTURBO.

Note: USB powered detection portion is added to board initialization
code so this feature is not supported in runtime. If AC adaptor is
unplugged after system boots-up then the behavior is still same, all
OPPs are supported.

Signed-off-by: AnilKumar Ch <anilkumar@ti.com>