ARM: OMAP2+: am335x: Correct pin-mux string for MMC0
GPIO pin was used for Card detect signal for MMC0 Instance. MMC
structure was updated previously but the pin-mux structure was left out.

There was no error seen because the pin-mux value was higher(7) than
mode found(6). When both were OR'ed, correct value was taken and the
error was covered. This patch corrects the same.

This patch applies both for EVM, Beaglebone.

Signed-off-by: Hebbar, Gururaja <gururaja.hebbar@ti.com>