ARM: OMAP2+: AM33XX: PM: Update the PER domain context loss counters when appropriate
In the current OMAP PM framework, device drivers depend on the
powerdomain context loss counters to save and restore the required
register context. In OMAP3 the counters associated with the context
loss are updated based on the LASTPOWERSTATEENTERED field and
PM_PREPWRSTST_<XYZ> registers. In OMAP4, the PRCM was updated
to provide context loss indication at a module level and this
has been handled appropriately.

AM335x devices neither have the PM_PREPWRSTST_<> registers nor do
they have module level context loss indicators. To woraround this
limitation, the context loss counters need to be updated in custom
code based on the knowledge of successful transition to a low power
state wherein the register context is lost. The current workaround
checks whether the peripheral powerdomain (PD_PER) successfully
transitioned to a state which leads to register context being lost
and updates the context loss counters appropriately.

While here, also replace the printks in the file with the appropriate
pr_*() versions.

Note: Due to the non-standard approach for updating the context loss
counters, this approach will most likely have to be revisited. Since
the powerdomain code has changed considerably in the mainline code,
reworking the current patch is being avoided for now.

Signed-off-by: Vaibhav Bedia <vaibhav.bedia@ti.com>