ARM: OMAP: AM33XX: PM: Different sleep sequence for DDR2 and DDR3
This patch enables having different suspend-resume sequences
for DDR2 and DDR3 in a single image.

Memory type and other ddr config values which are different for
DDR2 and DDR3 are passed to the sleep code from pm33xx.c.
Only two parameters are passed to assembly sleep code here;
memory type and suspend_vtp_control value.

Based on the memory type, two differnet suspend-resume sequences
are executed for DDR2 and DDR3 in the sleep file.

This has been tested on EVM, BBB and SK boards.

Signed-off-by: Satyanarayana, Sandhya <sandhya.satyanarayana@ti.com>