ARM: OMAP2+: am335x: platform data addition to add voltage ramp delay
This patch adds the minimum delay required to ramp the MPU voltage
from minimum OPP to maximum OPP, this is required only for BeagleBone.

MPU voltage transition from OPP50 to OPPTURBO took minimum of 70ms
to ramp completely when observed using an oscilloscope.

The reason for extended ramp time requirement on BeagleBone is not known.
This ramp time varies from board - board, if the board hangs with this
70msec delay also then try to increase this value.

Signed-off-by: AnilKumar Ch <anilkumar@ti.com>