drivers: net: ethernet: cpsw: dual emac remove module kernel crash bug fix
In dual emac, while removing the ti_cpsw module leads to kernel crash as the
second network interface created is not unregistered and freed. Adding support
to unregister and free the second network interface.

Do not set the second network interface to platform driver data as there is no
separate device for second network interface, so that in cpsw_remove() the
platform device received corresponds to the first network interface all the
time.

Signed-off-by: Mugunthan V N <mugunthanvnm@ti.com>