BeagleBone: Add board name for LCD7 cape lookup
The board name in the eeprom of some LCD7 capes doesn't match the one
in the lookup code of the board ("BeagleBone LCD7 CAPE" vs.
"BeagleBone LCD Cape"), also the strings in the eeprom of these
capes arent properly terminated. Check additionally whether the board
name starts with "BeagleBone LCD7 CAPE" to identify a LCD7 cape

Change-Id: I6df3bda794ef25504765907a375ce1700735a5e8
Signed-off-by: Matthias Kaehlcke <matthias@kaehlcke.net>
Signed-off-by: Vishveshwar Bhat <vishveshwar.bhat@ti.com>