AM335x: Provide the Actual min/max values of the Touch Screen
This patch adds the touch screen minimum and maximum values to the board file.
The ti_tscadc driver uses this platform data to specify the adc reading range.

Based on the patch submitted by Matthias Kaehlcke to Rowboat:
https://groups.google.com/d/topic/rowboat/sE_OqvfCfss/discussion

The Min/Max values are tuned and tested on the following board revisions:
  - AM335xEVM Rev. 1.1A and 1.2A.
  - Beaglebone LCD7 Cape Rev. A1 and A2.
For other revisions, if required the Min/Max values need to be tuned again.

Change-Id: I5880eb516a3cb749610bc329ecd0307e2b460dc3
Signed-off-by: Matthias Kaehlcke <matthias@kaehlcke.net>
Signed-off-by: Umakanta Patro <umakanta.patro@ti.com>