ARM: OMAP: AM33XX: PM: Fix potential deadlock in resume process due to MPU leaving GPIO0 enabled
Currently the GPIO driver is not fully runtime PM adapted
and the kernel leaves the module enabled during suspend.
On AM335x, there is a small window (maybe a few msecs)
wherein this causes problems with suspend-resume.

After the MPU has executed WFI and while the Cortex-M3 is
running, any GPIO0 activity comes as a normal interrupt to
the nVIC of Cortex-M3. However, Cortex-M3 is meant to handle
only the wakeup interrupts (sWAKEUP) from the various wakeup
peripherals so the normal interrupt from GPIO0 bank goes
unhandled. This leads to the GPIO0 module not generating
a wakeup interrupt on repeated events until a resume happens
via a different wakeup source (say UART) and the MPU
services the regular GPIO interrupt which came in during
the suspend process.

On systems where the only wakeup source is over GPIO0, this
can potentially lead to a deadlock situation.

To avoid this situation, disable the GPIO0 module from
MPU as part of the suspend sequence. Once this is done,
any activity of the GPIO0 bank gets registered as a wakeup
interrupt in the nVIC of Cortex-M3. After Cortex-M3 runs
through its suspend sequence, due to the pending wakeup
interrupt from GPIO0 the system resumes and thereby
a deadlock as described above is avoided.

Note:- Once the driver is fully adapted to runtime PM,
the PM framework in the kernel is expected to handle
the disable of the GPIO banks and hence this workaround
will not be needed.

Signed-off-by: Vaibhav Bedia <vaibhav.bedia@ti.com>

merge with the previous