Fix cold boot functionality issue with android gadget
Android gadget requires the pullup to be low untill
userland comes up, configures the usb functions and
then enables the pullup.

By default, the udc-core layer enables the pullup
during gadget_probe_driver, this causes break in
the cold boot enumeration on the Host for android.

This patch keeps the udc-core pullup if CONFIG_USB_G_ANDROID
is not enabled.

This patch fixes the "Unknown Device" popup issue
on Windows host if booted up with usb connected.

Change-Id: I845698a3cad8bd22dc9f16548dc1d748a5db6a01
Signed-off-by: Moiz Sonasath <m-sonasath@ti.com>
Signed-off-by: Arun Joseph <arunjoseph@ti.com>