usb: musb: ti81xx: fix gadget enumerated as Fullspeed device
When musb is in otg/peripheral mode, no matter the gadget driver is
built in kernel or as a module, if connected to USB host before power on
the platform, the host enumerates the platform as Fullspeed device
after powered it on.

The root cause is that before the gadget driver is initialized,
otg_timer() toggles SESSION bit in devctl register which enables D+/D-
pullup, which causes host PC starts to enumerate before the gadget
driver notifying it is a Highspeed device.

The solution is that do not toggle the SESSION bit if VBUS presents,
which means already connected to a host.

Signed-off-by: Bin Liu <b-liu@ti.com>
Signed-off-by: Ravi Babu <ravibabu@ti.com>