ARM: OMAP: AM33XX: PM: Update sleep sequence for PG2.0
PG2.0 has bug fix for advisory 1.0.17 because of which we can now
write to REG_PHY_ENABLE_DYNAMIC_PWRDN to disable the input buffers.
This is done in u-boot as part of DDR configurations.

On enabling Dynamic PowerDown in PG2.0, the below workarounds can
be avoided in suspend sequence.
- switching to LVCMOS mode (setting MDDR_SEL in DDR_IO_CTRL)
- configuring pull downs on DATA0 and DATA1 macros
(As explained by Hardware Apps team)

In this patch, switching to LVCMOS mode is skipped for PG2.0.

However, configuring pulldowns on DATA0 and DATA1 macros is
retained since we measured a higher number for VDDS_DDR when this
step was removed.

This has been tested on EVM1.5,EVM-1.2a,EVM-SK1.2b,BBB.

Signed-off-by: Satyanarayana Sandhya <sandhya.satyanarayana@ti.com>