video: nxp: tda998x: driver for nxp tda998x HDMI chip
This is a backport of the NXP TDA998X i2c encoder slave
present in 3.8 kernel.

This driver still make use of some of the DRM definitions,
but works with the DA8xx framebuffer driver in 3.2 kernel.

This driver has been validated for three modes -
1280x720(CEA-4), 1024x768 and 640x480(CEA-1)
Tested on HDMI compliant TVs - Sony, Samsung and Lilliput.

By default the driver enables 1280x720(CEA-4), this can be
changed by providing a different DRM mode in the driver source.
Check drm_edid_modes.h for a list of DRM modes. But make sure
corresponding display configuration is chosen for the framebuffer
driver.

The resolution can't be changed after boot.

Change-Id: I39cc96f889777b2bbfff2d2d454d3ab903dda901
Signed-off-by: Arun Joseph <arunjoseph@ti.com>