bq27xxx: update the locking mechanism
Currently the mutex is being locked only at
one place. It's not clear what it is locking
against. It doesn't seem to serve any purpose.

Instead of the current scheme, use di->lock to
ensure mutual execution between differnt threads
of execution so that the commands on the bus are
not mixed from the two threads. The two threads
of execution in the driver are:

1. get_property() calls from the power supply class
2. bq27x00_battery_poll() called from a delayed work
   thread

Signed-off-by: Aneesh V <aneesh@ti.com>