power: bq27xxx: add support for DM initialization of ROM based gauges
ROM based gauges such as bq274xx series do not have data flash.
Instead the configuration parameters are copied from ROM to volatile
dataram or data memory(DM) on startup of the gauge. Any customization
to the default parameters should be done by system software. After
a POR or a Control() RESET command the DM returns to the ROM defaults.
The customizations have to be re-configured by the system software.

Signed-off-by: Aneesh V <aneesh@ti.com>